+++ src/lightroom/__init__.py
@@ -32,7 +32,7 @@ from .exceptions import (
     PluginHandshakeError,
 )
 
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 
 __all__ = [
     "LightroomClient",